Move func to xtime pkg
