+++ xtime/parse_format_name.go
@@ -1,4 +1,4 @@
-package util
+package xtime
 
 import (
 	"fmt"
